Refactor search functionality to use SearchColumnProviderHook for custom variable retrieval and added new SearchColumnProviderHook class
