+++ library/Icingadb/Web/Control/SearchBar/ObjectSuggestions.php
@@ -9,7 +9,7 @@ use Icinga\Module\Icingadb\Common\Auth;
 use Icinga\Module\Icingadb\Common\Database;
 use Icinga\Module\Icingadb\Data\QueryColumnsProvider;
 use Icinga\Module\Icingadb\Data\QueryValuesProvider;
-use Icinga\Module\Icingadb\Hook\CustomVarsRetrieverHook;
+use Icinga\Module\Icingadb\Hook\SearchColumnsProviderHook;
 use ipl\Html\HtmlElement;
 use ipl\Orm\Model;
 use ipl\Stdlib\BaseFilter;
@@ -93,8 +93,8 @@ class ObjectSuggestions extends Suggestions
         $resolver = $model::on($this->getDb())->getResolver();
 
         $quickFilter = Filter::any();
-        $customVarColumns = CustomVarsRetrieverHook::getCustomVarColumns($model);
-        $columns = [...$model->getSearchColumns(), ...$customVarColumns];
+        $defaultColumns = $model->getSearchColumns();
+        $columns = SearchColumnsProviderHook::getCustomVarColumns($model, $defaultColumns);
 
         foreach ($columns as $column) {
             if (strpos($column, '.') === false) {
