+++ library/Icingadb/Web/Controller.php
@@ -22,6 +22,7 @@ use Icinga\Module\Icingadb\Data\CsvResultSet;
 use Icinga\Module\Icingadb\Data\JsonResultSet;
 use Icinga\Module\Icingadb\Data\QueryColumnsProvider;
 use Icinga\Module\Icingadb\Web\Control\ColumnChooser;
+use Icinga\Module\Icingadb\Hook\SearchColumnsProviderHook;
 use Icinga\Module\Icingadb\Web\Control\GridViewModeSwitcher;
 use Icinga\Module\Icingadb\Web\Control\TimestampToggle;
 use Icinga\Module\Icingadb\Web\Control\ViewModeSwitcher;
@@ -349,8 +350,8 @@ class Controller extends CompatController
     protected function prepareSearchFilter(Query $query, string $search, Filter\Any $filter, array $additionalColumns)
     {
         $model = $query->getModel();
-        $customVarColumns = CustomVarsRetrieverHook::getCustomVarColumns($model);
-        $columns = array_merge($model->getSearchColumns(), $additionalColumns, $customVarColumns);
+        $defaultColumns = $model->getSearchColumns();
+        $columns = SearchColumnsProviderHook::getCustomVarColumns($model, $defaultColumns);
         foreach ($columns as $column) {
             if (strpos($column, '.') === false) {
                 $column = $query->getResolver()->qualifyColumn($column, $query->getModel()->getTableName());
